+++ src/commands/deployWorkspaceProject/deploymentConfiguration/tree/getTreeItemDeploymentConfiguration.ts
@@ -3,8 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { type IContainerAppContext } from "../../IContainerAppContext";
-import { type DeploymentConfiguration } from "./DeploymentConfiguration";
+import { type IContainerAppContext } from "../../../IContainerAppContext";
+import { type DeploymentConfiguration } from "../DeploymentConfiguration";
 
 // Todo: Monorepo core logic (tree item path) https://github.com/microsoft/vscode-azurecontainerapps/issues/613
 export async function getTreeItemDeploymentConfiguration(_: IContainerAppContext): Promise<DeploymentConfiguration> {
